[GIT] Migrate training data to Git LFS
The training data is too large to be committed to Github. Git Large File
Storage (LFS) replaces large files with text pointers and stores the
actual contents of the file on a remote server. This allows the massive
training data files to be committed without violating the Github's
oversize file policy.