[ADD] Markov and Random Engines
Modifies the engine interface; chess engines should select an optimal move from a set of candidates moves. Adds a simple Random Engine that selects the optimal move uniformly at random from the set of candidate moves. This engine is a useful benchmark for measuring the performance of more complicated chess engines. Adds an incomplete implementation of a Markov Engine that utilizes a Markov-Chain to determine the optimal move.